close #3: using int64 to avoid conversion from uint64 to int64, int64 is sufficient for a binlog file.

+++ get_gtid_of_binlog.go
@@ -16,12 +16,12 @@ func GetGtidOfBinlog(binlogPath string) (gtidDesc string, err error) {
 	}
 	defer file.Close()
 
-	p := uint64(4)
+	p := int64(4)
 	headerBs := make([]byte, 19)
 	gtidBs := make([]byte, 25)
 
 	for {
-		if _, err := file.Seek(int64(p), 0); nil != err {
+		if _, err := file.Seek(p, 0); nil != err {
 			if "EOF" == err.Error() {
 				break
 			}
@@ -39,7 +39,7 @@ func GetGtidOfBinlog(binlogPath string) (gtidDesc string, err error) {
 		eventType := int(headerBs[4])
 
 		if GTID_LOG_EVENT != eventType {
-			p += uint64(length)
+			p += int64(length)
 			continue
 		}
 
@@ -56,7 +56,7 @@ func GetGtidOfBinlog(binlogPath string) (gtidDesc string, err error) {
 			return gtidDesc, err
 		}
 
-		p += uint64(length)
+		p += int64(length)
 	}
 	return gtidDesc, nil
 }
